removed unnecessary require statements, support/ is auto loaded
