+++ UPtables/features/support/env.rb
@@ -7,10 +7,8 @@
 require 'cucumber/rails'
 require 'factory_girl_rails'
 require 'capybara-screenshot/cucumber'
-require_relative 'helper'
 
 World(FactoryGirl::Syntax::Methods)
-World(Helper)
 
 Capybara.javascript_driver = :webkit
